refactor: domain directory
